controllerの名前を変更、views以下のファイルの配置を変更
