+++ bbsApp/controllers/indexController.js
@@ -1,0 +1,7 @@
+const { validationResult } = require('express-validator');
+
+module.exports = {
+  goIndex(req, res) {
+    res.render('index', {title: 'Index'});
+  }
+}
