--- bbsApp/views/allPosts.ejs
@@ -1,4 +1,0 @@
-<%- include('components/header'); %>
-<%- include('components/navbar_logged_in'); %>
-<h1>みんなの投稿</h1>
-<%- include('components/footer'); %>
