+++ bbsApp/views/components/navbar.ejs
@@ -9,16 +9,16 @@
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
       <ul class="navbar-nav ms-auto mb-2 mb-lg-0">
         <li class="nav-item">
-          <a class="nav-link" href="/allPosts">みんなの投稿</a>
+          <a class="nav-link" href="/user/allPosts">みんなの投稿</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="/myPost">投稿を作成</a>
+          <a class="nav-link" href="/user/myPost">投稿を作成</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="/login">Login</a>
+          <a class="nav-link" href="/auth/login">Login</a>
         </li>
         <li class="nav-item">
-          <a class="nav-link" href="/register">Register</a>
+          <a class="nav-link" href="/auth/register">Register</a>
         </li>
       </ul>
     </div>
