--- bbsApp/views/myPost.ejs
@@ -1,4 +1,0 @@
-<%- include('components/header'); %>
-<%- include('components/navbar_logged_in'); %>
-<h1>投稿を作成</h1>
-<%- include('components/footer'); %>
